Correct calculation of propensities in example
Propensities had not been calculated correctly when rules infered structurally
equivalent reactions. Consider the the LigationRule k + l --> k.l. If k='aa'
and l='b' this rule infers two Reactions (for products 'aab' and 'baa')
with stoichiometric rate constant c. If k='aa' and l='a' this rule infers a
single reaction (with product 'aaa'). To compensate for this, the
stoichiometric rate constant has to be set to 2*c.

This leads to rather complex code in the inference methods. An alternative
would be to simply infer the reaction a+aa-->aaa twice and have stocal
deal with the multiplicity. This however forces the user to exactly infer
all newly applicable reactions. (Currently it was okay to infer reactions
that where previously active already).