Allow for time-dependent reaction rates
Reaction.propensity can now accept an additional argument time.
If time is given in the propensity signature, the reaction rate
is assumed to be time dependent and the Reaction property
is_autonomous evaluates to False.
Reaction.propensity_integral performs numerical integration of
the propensity function over a given time window, and
Reaction.propensity_meets_target calculates the size of the
time window for which the propensity integral meets a given value.
Both methods should be overridden if analytic expressions are known
for these equations.