Fix issue #3
AndersonNRM held the next occurrence "clocks" T and P only local
variables of __iter__. This did not allow to add new clocks for
inferred transitions. Now T and P are proper attributes of sampler
instances that are updated in add_transition.

This closes bug #3 and the behavior to fire inferred transitions
(events) is now enforced in test_algorithms.TestFirstReactionMethod