Fix bug in TestFirstReactionMethod specification
TestFirstReactionMethod.test_exact_number_of_events actually
miscounted the number of events that should occurr (10 instead of 11).
The wrong specification lead to a behavior that events that occurr
exactly at tmax are not executed. I have changed both specification
as well as the implementation of TrajectorySampler.__iter__ to *include*
transitions occurring at tmax.