Fix yet another AndersonNRM issue
When advancing internal clocks Tk of affected reactions, I advanced
them for the time interval from self.time to (the next firing) time.
However, internal clocks need to be advanced for the interval starting
with their last update.

I now keep track of the global time that a reaction was updated as
another data item of each transition (data.t). The update interval
is now calculated as time-data.t.