Add towards public documentation
Public API started in doc/api.rst. This file and its future decendents
will serve as semantic versioning specification.

An implementation for the Trajectory class is provided by not yet
tested and its interface is not fixed in the spec yet.

The commit also includes examples of Jupyter notebook sessions.